requirements.txt
streamlit
python-docx
openai>=1.40.0
python-dotenv
pydantic
pandas
pytest
httpx>=0.24.0
